changed get_sprites by get_elem_into_list to make it generic, prototype now in the map header
